Optimize packet size and fix pcnt overflow issues

+++ arduino/main/globals.h
@@ -21,7 +21,7 @@ extern QueueHandle_t commandQueue;
 extern pcnt_unit_t pcnt_unit_left;
 extern pcnt_unit_t pcnt_unit_right;
 extern std::atomic<uint8_t> pageIndex;
-extern std::atomic<float> distanceFront;
+extern std::atomic<int16_t> distanceFront;
 extern std::atomic<uint8_t> pendingPIDMode; // 0 = PID control mode, 1 = open-loop PWM control mode
 
 // RTOS task handles
